Create readme with disclaimer
Users should be aware of the risks of using this software. This software has been created for personal use, not keeping in mind that someone unfamiliar with its limitations and shortcomings will be using it and it should be made clear to anyone using it that functions may be buggy, half-implemented, and may not work as expected. I've added a disclaimer in an attempt to make this clear.